配置锁定状态接口

+++ foxconn-sw-tool-data/src/main/java/com/foxconn/sw/data/dto/entity/oa/WorkReportDetail.java
@@ -6,9 +6,9 @@ public class WorkReportDetail {
     private int num;
     private String project;
     private String description;
-    private int day;
-    private int target;
-    private int current;
+    private Double day;
+    private Integer target;
+    private Integer current;
     private String remark;
 
     public int getId() {
@@ -43,27 +43,27 @@ public class WorkReportDetail {
         this.description = description;
     }
 
-    public int getDay() {
+    public Double getDay() {
         return day;
     }
 
-    public void setDay(int day) {
+    public void setDay(Double day) {
         this.day = day;
     }
 
-    public int getTarget() {
+    public Integer getTarget() {
         return target;
     }
 
-    public void setTarget(int target) {
+    public void setTarget(Integer target) {
         this.target = target;
     }
 
-    public int getCurrent() {
+    public Integer getCurrent() {
         return current;
     }
 
-    public void setCurrent(int current) {
+    public void setCurrent(Integer current) {
         this.current = current;
     }
