+++ foxconn-sw-tool-data/src/main/java/com/foxconn/sw/data/dto/entity/oa/WorkReportParams.java
@@ -11,7 +11,7 @@ public class WorkReportParams {
     private String project;
     @NotNull(message = "内容不能为空")
     private String description;
-    private Integer day;
+    private Double day;
     private Integer target;
     private Integer current;
     private String remark;
@@ -57,11 +57,11 @@ public class WorkReportParams {
         this.description = description;
     }
 
-    public Integer getDay() {
+    public Double getDay() {
         return day;
     }
 
-    public void setDay(Integer day) {
+    public void setDay(Double day) {
         this.day = day;
     }
